# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v4/group:
    post:
      operationId: GroupCreate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Function"
      responses:
        "201":
          description: Create group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDto"
        "401": &a17
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AxiosErrorDto"
              example:
                response:
                  data:
                    statusCode: 401
                    error: Unauthorized
                    message: Human readable error
                status: 401
        "403": &a18
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AxiosErrorDto"
              example:
                response:
                  data:
                    statusCode: 403
                    error: Forbidden
                    message: Human readable error
                status: 403
        "500": &a19
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AxiosErrorDto"
              example:
                response:
                  data:
                    statusCode: 500
                    error: Internal Server Error
                    message: Human readable error
                status: 500
      security: &a20
        - bearer: []
        - jwt: []
      tags: &a21
        - group
    get:
      operationId: GroupFindAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: perPage
          required: false
          in: query
          schema:
            example: 10
            type: number
      responses:
        "200":
          description: Get all groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupListItemDto"
        "401": *a17
        "403": *a18
        "500": *a19
      security: *a20
      tags: *a21
  /api/v4/group/{id}:
    get:
      operationId: GroupFindOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Get one group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDto"
        "401": *a17
        "403": *a18
        "500": *a19
      security: *a20
      tags: *a21
    put:
      operationId: GroupUpdate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Function"
      responses:
        "200":
          description: Update group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDto"
        "401": *a17
        "403": *a18
        "500": *a19
      security: *a20
      tags: *a21
    delete:
      operationId: GroupRemove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "204":
          description: Delete group
        "401": *a17
        "403": *a18
        "500": *a19
      security: *a20
      tags: *a21
info:
  title: PÉK API
  description: Profiles and Groups
  version: v4
  contact:
    name: kir-dev
    url: https://kir-dev.hu
    email: hello@kir-dev.hu
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    cookie:
      type: apiKey
      in: cookie
      name: jwt
  schemas:
    UnauthorizedErrorDto:
      type: object
      properties:
        message:
          type: string
          default: JWT cookie or Bearer token not found
          enum:
            - Authorization token not found
        statusCode:
          type: number
          default: 401
          enum:
            - 401
        error:
          type: string
          default: Unauthorized
          enum:
            - Unauthorized
          nullable: true
      required:
        - message
        - statusCode
    InternalServerErrorDto:
      type: object
      properties:
        statusCode:
          type: number
          default: 500
          enum:
            - 500
        message:
          type: string
          default: Internal Server Error
          enum:
            - Internal Server Error
      required:
        - statusCode
        - message
    ForbiddenErrorDto:
      type: object
      properties:
        statusCode:
          type: number
          default: 403
          enum:
            - 403
        message:
          type: string
        resourceId:
          type: string
          example: pek:group:123:profile
          nullable: true
        resourceOp:
          type: string
          enum:
            - CREATE
            - READ
            - UPDATE
            - DELETE
          nullable: true
        error:
          type: string
          enum:
            - Forbidden
          nullable: true
      required:
        - statusCode
        - message
    AxiosErrorResponseDto:
      type: object
      properties:
        data:
          description: Data of type T
          oneOf:
            - $ref: "#/components/schemas/InternalServerErrorDto"
            - $ref: "#/components/schemas/UnauthorizedErrorDto"
            - $ref: "#/components/schemas/ForbiddenErrorDto"
      required:
        - data
    AxiosErrorDto:
      type: object
      properties:
        response:
          $ref: "#/components/schemas/AxiosErrorResponseDto"
        status:
          type: number
      required:
        - response
        - status
...
    Function:
      type: object
      properties: {}
    GroupListItemDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the group
          example: cjld2cjxh0000qzrmn831i7rn
        name:
          type: string
          description: Unique name of the group
          example: Developers
      required:
        - id
        - name
    MemberUserDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the user
          example: cjld2cjxh0000qzrmn831i7rn
        firstName:
          type: string
          description: First name of the member
          example: John
        lastName:
          type: string
          description: Last name of the member
          example: Doe
        nickname:
          type: string
          description: Nickname of the member
          example: Johnny
      required:
        - id
        - firstName
        - lastName
        - nickname
    MemberListItemDto:
      type: object
      properties:
        user:
          description: User profile of the member
          allOf:
            - $ref: "#/components/schemas/MemberUserDto"
      required:
        - user
    GroupDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the group
          example: cjld2cjxh0000qzrmn831i7rn
        name:
          type: string
          description: Unique name of the group
          example: Developers
        description:
          type: string
          description: Description of the group purpose and activities
          example: Group for software developers
        purpose:
          type: string
          description: The primary purpose/type of the group
          enum:
            - UNKNOWN
            - OLD
            - COMMITTEE
            - PARTY
            - CIRCLE
            - D
            - ELLIPSE
            - YEAR_CLASS
            - GROUP
            - CULTURE
            - PROJECT
            - EVENT
            - RESORT
            - SPORT
            - PROFESSIONAL
            - FLOOR
            - SERVICE
          example: GROUP
        isCommunity:
          type: boolean
          description: Whether this group is a community
          example: true
        isResort:
          type: boolean
          description: Whether this group is a resort
          example: false
        isTaskForce:
          type: boolean
          description: Whether this group is a task force
          example: false
        hasTransitiveMembership:
          type: boolean
          description: Whether this group inherits members from child groups
          example: true
        children:
          description: Children groups of this group
          type: array
          items:
            $ref: "#/components/schemas/GroupListItemDto"
        parent:
          nullable: true
          description: Parent group of this group
          allOf:
            - $ref: "#/components/schemas/GroupListItemDto"
        memberships:
          description: Members of this group
          type: array
          items:
            $ref: "#/components/schemas/MemberListItemDto"
      required:
        - id
        - name
        - description
        - purpose
        - isCommunity
        - isResort
        - isTaskForce
        - hasTransitiveMembership
        - children
        - parent
        - memberships
...